Blazor desktop templates with Router and Layout components
