+++ src/BlazorWinForms/BlazorWinForms/.template.config/dotnetcli.host.json
@@ -1,0 +1,3 @@
+{
+  "$schema": "http://json.schemastore.org/dotnetcli.host"
+}
