+++ src/BlazorWinForms/BlazorWinForms/.template.config/ide.host.json
