+++ src/BlazorWinForms/BlazorWinForms/.template.config/template.json
@@ -1,19 +1,18 @@
 {
     "$schema": "https://json.schemastore.org/template",
     "author": "Vijay Anand E G",
-    "defaultName": "BlazorWinForms1",
+    "defaultName": "BlazorWinForms",
     "classifications": [
         "Windows",
         "Desktop",
-        "WinForms",
         "Windows Forms",
         "Blazor"
     ],
     "identity": "VijayAnand.WindowsForms.Blazor",
     "description": "A project for hosting Blazor component in the Windows Forms desktop application",
-    "name": "Windows Forms with Blazor component (.NET 6 Preview 7)",
+    "name": "Windows Forms App with Blazor component (.NET 6 Preview 7)",
     "shortName": "winforms-blazor",
-    "sourceName": "BlazorWinForms1",
+    "sourceName": "BlazorWinForms",
     "tags": {
         "language": "C#",
         "type": "project"
