+++ src/BlazorWinForms/BlazorWinForms/App.razor
@@ -1,0 +1,10 @@
+﻿<Router AppAssembly="@GetType().Assembly">
+	<Found Context="routeData">
+		<RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
+	</Found>
+	<NotFound>
+		<LayoutView Layout="@typeof(MainLayout)">
+			<p>Sorry, there's nothing at this address.</p>
+		</LayoutView>
+	</NotFound>
+</Router>
