+++ src/BlazorWinForms/BlazorWinForms/Index.razor
@@ -1,0 +1,3 @@
+﻿@page "/"
+
+<Counter />
