+++ src/BlazorWinForms/BlazorWinForms/MainLayout.razor
@@ -1,0 +1,5 @@
+﻿@inherits LayoutComponentBase
+
+<div>
+    @Body
+</div>
