+++ src/BlazorWinForms/BlazorWinForms/Program.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace BlazorWinForms1
+namespace BlazorWinForms
 {
     static class Program
     {
