+++ src/BlazorWinForms/Install-Template.bat
@@ -1,6 +1,6 @@
 :: Installs the NuGet package
 @echo off
 echo Installing the project template . . .
-dotnet new --install .\bin\Debug\VijayAnand.WindowsForms.Blazor.1.0.3.nupkg
+dotnet new --install .\bin\Debug\VijayAnand.WindowsForms.Blazor.1.0.4.nupkg
 echo Process completed
 pause
