+++ src/BlazorWpf/BlazorWpf/.template.config/ide.host.json
