+++ src/BlazorWpf/BlazorWpf/.template.config/template.json
@@ -1,7 +1,7 @@
 {
     "$schema": "https://json.schemastore.org/template",
     "author": "Vijay Anand E G",
-    "defaultName": "BlazorWpf1",
+    "defaultName": "BlazorWpf",
     "classifications": [
         "Windows",
         "Desktop",
@@ -10,9 +10,9 @@
     ],
     "identity": "VijayAnand.Wpf.Blazor",
     "description": "A project for hosting Blazor component in the WPF desktop application",
-    "name": "WPF with Blazor component (.NET 6 Preview 7)",
+    "name": "WPF App with Blazor component (.NET 6 Preview 7)",
     "shortName": "wpf-blazor",
-    "sourceName": "BlazorWpf1",
+    "sourceName": "BlazorWpf",
     "tags": {
         "language": "C#",
         "type": "project"
