+++ src/BlazorWpf/BlazorWpf/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="BlazorWpf1.App"
+﻿<Application x:Class="BlazorWpf.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:BlazorWpf1"
+             xmlns:local="clr-namespace:BlazorWpf"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
