+++ src/BlazorWpf/BlazorWpf/AssemblyInfo.cs
