+++ src/BlazorWpf/BlazorWpf/Counter.razor
@@ -1,4 +1,4 @@
-@using Microsoft.AspNetCore.Components.Web
+@page "/counter"
 
 <h1>Counter</h1>
