+++ src/BlazorWpf/BlazorWpf/Index.razor
@@ -1,0 +1,3 @@
+﻿@page "/"
+
+<Counter />
