+++ src/BlazorWpf/BlazorWpf/MainWindow.xaml
@@ -1,10 +1,10 @@
-﻿<Window x:Class="BlazorWpf1.MainWindow"
+﻿<Window x:Class="BlazorWpf.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:blazor="clr-namespace:Microsoft.AspNetCore.Components.WebView.Wpf;assembly=Microsoft.AspNetCore.Components.WebView.Wpf"
-        xmlns:local="clr-namespace:BlazorWpf1"
+        xmlns:local="clr-namespace:BlazorWpf"
         mc:Ignorable="d"
         Title="MainWindow"
         Height="450"
@@ -13,7 +13,7 @@
 		<blazor:BlazorWebView HostPage="wwwroot/index.html"
 		                      Services="{StaticResource services}">
 			<blazor:BlazorWebView.RootComponents>
-				<blazor:RootComponent ComponentType="{x:Type local:Counter}"
+				<blazor:RootComponent ComponentType="{x:Type local:Main}"
 				                      Selector="#app"/>
 			</blazor:BlazorWebView.RootComponents>
 		</blazor:BlazorWebView>
