+++ src/BlazorWpf/BlazorWpf/_Imports.razor
@@ -1,0 +1,3 @@
+﻿@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.AspNetCore.Components.Web
+@using BlazorWpf
