+++ src/BlazorWpf/BlazorWpf/wwwroot/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
     <title>Blazor app</title>
     <base href="/" />
-    <link href="BlazorWpf1.styles.css" rel="stylesheet" />
+    <link href="BlazorWpf.styles.css" rel="stylesheet" />
     <link href="css/app.css" rel="stylesheet" />
 </head>
