+++ src/BlazorWpf/Install-Release-Template.bat
@@ -1,6 +1,6 @@
 :: Installs the NuGet package
 @echo off
 echo Installing the project template . . .
-dotnet new --install .\bin\Debug\VijayAnand.Wpf.Blazor.1.0.5.nupkg
+dotnet new --install .\bin\Release\VijayAnand.Wpf.Blazor.1.0.5.nupkg
 echo Process completed
 pause
