+++ src/BlazorWpf/release-notes.txt
@@ -1,3 +1,5 @@
-Project template updated with .NET 6 Preview 7 package dependencies.
+What's new in ver. 1.0.5:
+
+Added Router and Layout components to the template.
 
 Note: From .NET 6 Preview 5 onwards, the target version of the WPF application has to be Windows 10 [Version 10.0.19041] (2004 / 20H1) for the BlazorWebView package to restore and compile.
